RIDER-26362: introduce a new folder mode for the common file dialogs
This improves UX for common file dialogs when the user wants to choose
the current folder. From now on, they'll be able to click "Select
Folder" button in cases when a folder is open, but no item is selected.

This new behavior is more similar to the FOS_PICKFOLDERS mode of the
common dialogs.